+++ README.md
@@ -1,7 +1,7 @@
-# porter
+# privateer
 
-[![PyPI - Version](https://img.shields.io/pypi/v/porter.svg)](https://pypi.org/project/porter)
-[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/porter.svg)](https://pypi.org/project/porter)
+[![PyPI - Version](https://img.shields.io/pypi/v/privateer.svg)](https://pypi.org/project/privateer)
+[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/privateer.svg)](https://pypi.org/project/privateer)
 
 -----
 
@@ -15,17 +15,17 @@
 ## Usage
 
 ```Usage:
-  porter --version
-  porter backup <path> --to=HOST [--exclude=TARGETS] [--include=INCLUDE]
-  porter restore <path> --from=HOST [--exclude=TARGETS] [--include=INCLUDE]
+  privateer --version
+  privateer backup <path> --to=HOST [--exclude=TARGETS] [--include=INCLUDE]
+  privateer restore <path> --from=HOST [--exclude=TARGETS] [--include=INCLUDE]
 
 Options:
   --exclude=TARGETS  Comma separated string of target names to exclude (default is to include all)
   --include=TARGETS  Comma separated string of target names to include (default is to include all)
 ```
 
-`<path>` is the path to a directory containing a `porter.json` file. This file should contain at least one target 
-and at least one host. See `./config/porter.json` for an example. By default all targets in the config file are used, 
+`<path>` is the path to a directory containing a `privateer.json` file. This file should contain at least one target 
+and at least one host. See `./config/privateer.json` for an example. By default all targets in the config file are used, 
 but this can be overridden by explicitly including or excluding targets by name.
 
 ## Test and lint
@@ -39,9 +39,9 @@ On CI, use `hatch run cov-ci` to generate an xml report.
 ## Installation
 
 ```console
-pip install porter
+pip install privateer
 ```
 
 ## License
 
-`porter` is distributed under the terms of the [MIT](https://spdx.org/licenses/MIT.html) license.
+`privateer` is distributed under the terms of the [MIT](https://spdx.org/licenses/MIT.html) license.
