+++ config/invalid/privateer.json
